Errors :ear: :tiger:

+++ app/controllers/api/v1/users_controller.rb
@@ -1,33 +1,44 @@
 class Api::V1::UsersController < ApplicationController
-  def index
-
-    render json: User.all
-  end
+  # def index
+  #   render json: User.all
+  # end
 
   def show
-
-    render json: User.find(params[:id])
+    user = User.find(params[:id])
+    if user == current_user
+      render json: user
+    else
+      render nothing: true
+    end
   end
 
   def create
-
-
     user = User.new(user_params)
-    user.save
-    render json: user
+    if user.save
+      render json: user
+    else
+      render json: {errors: user.errors}, status: :unprocessable_entity
+    end
   end
 
   def update
-
     user = User.find(params[:id])
-    user.update(user_params)
-    render json: user
+    if user == current_user
+      if user.update(user_params)
+        render json: user
+      else
+        render json: {errors: user.errors}, status: :unprocessable_entity
+      end
+    else
+      render nothing: true
+    end
   end
 
   def destroy
-
     user = User.find(params[:id])
-    user.destroy
+    if user == current_user
+      user.destroy
+    end
     render nothing: true
   end
